+++ packages/data-explorer/src/components/dataView/TableRow.vue
@@ -98,8 +98,13 @@ export default {
     position: -webkit-sticky; /* for Safari */
     position: sticky;
     left: 0;
+<<<<<<< HEAD
     background-color: var( --gray-light );
     border-right: 1px var( --border ) solid;
+=======
+    background-color: var(--light);
+    border-right: 1px var(--border) solid;
+>>>>>>> master
   }
   /**
    * Fix borders in the first column
@@ -124,7 +129,7 @@ export default {
   * Style row actions
   */
   .mg-data-column {
-    max-width: 8rem;
+    max-width: 32rem;
     overflow: hidden;
   }
   [role="button"]{